import of 6.1.1 version

+++ viewer/src/org/icepdf/ri/common/PrintHelper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2014 ICEsoft Technologies Inc.
+ * Copyright 2006-2016 ICEsoft Technologies Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the
@@ -47,7 +47,6 @@ public class PrintHelper implements Printable {
     private PageTree pageTree;
     private Container container;
     private float userRotation;
-
     private boolean printFitToMargin;
     private int printingCurrentPage;
     private int totalPagesToPrint;
@@ -350,8 +349,8 @@ public class PrintHelper implements Printable {
         PDimension pageDim = currentPage.getSize(userRotation);
 
         // Grab default page width and height
-        float pageWidth = pageDim.getWidth();
-        float pageHeight = pageDim.getHeight();
+        float pageWidth = (float) pageDim.getWidth();
+        float pageHeight = (float) pageDim.getHeight();
 
         // Default zoom factor
         float zoomFactor = 1.0f;
